requirements.txt
# NBA FanDuel Simulator Dependencies

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0

# Configuration
PyYAML>=6.0

# Optional: for visualization (not required for core functionality)
matplotlib>=3.7.0
seaborn>=0.12.0
